Beginnings of a podcast API call
- /api/podcast returns a list of podcasts that the server is subscribed
to.

- /api/podcast/$PODCASTID returns a list of the associated podcast's
stored episodes

-
/api/podcast?action=add&podcastUrl=$URLENCODEDURL&podcastKeepCap=$CUSTOM
KEEPCAP adds a podcast subscription to the server and downloads the
specified number of episodes

- For now, the keep cap is mandatory, but in the future, if it is not
supplied, a default keep cap will be used, which will be specified in
the config file.

- Renamed the Podcast namespace to PodcastManagement in order to avoid
the collision between the Podcast class and the Podcast namespace

- Podcast directory creation is now deferred until new episodes are
downloaded rather than when a Podcast object is created.