Fixed a couple audio transcoding issues
- Fixed some logging that was assuming a certain key existed in a
dictionary when it wouldn't necessarily, which caused a crash

- Added a return statement in AbstractTranscoder to prevent a crash if
ffmpeg wasn't correctly installed

- Removed the quotes surrounding the OutputPath return string.  We
assemble this path.  It will never have any characters that need
escaping.  This was preventing the runtime from finding the file, as it
was thinking that the file name should have quotes, which is impossible.

- Somehow we overlooked the fact that the Song data model object should
return ItemType.Song instead of its super's type.