- Replaced SuperWebSocket with SignalR (uses long polling for better reliability with http proxy servers and better browser support) - Implemented skeleton class for device syncing - Implemented javascript skeleton functions for device syncing on the client side - Added test html file to verify communication
(note: RawConnection.cs is likely not necessary and will be removed once we determine it's not being used. I read something that indicated it may be used as a fallback, but I'm not sure that was correct. Either way I'd rather have it in git just in case, we can remove later)